Updates to Order Functionality
