+++ backend/sep2025_project_team_004/sep2025_project_team_004/store/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import ProductListView, ReviewCreateAPIView, CreateOrderView
+from .views import ProductListView, ReviewCreateAPIView, CreateOrderView, MyOrdersListView
 
 app_name = "store" 
 
@@ -7,4 +7,5 @@ urlpatterns = [
     path("products/", ProductListView.as_view(), name="product-list"),
     path('reviews/', ReviewCreateAPIView.as_view(), name='review-create'),
     path("orders/create/", CreateOrderView.as_view(), name="create-order"),
+    path("orders/my/", MyOrdersListView.as_view(), name="my-orders"),
 ]
