+++ mobile/app/my-orders.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import {
   View,
   Text,
@@ -8,46 +8,162 @@ import {
   StyleSheet,
   TouchableOpacity,
   Dimensions,
+  Modal,
+  TextInput,
+  Alert,
+  ActivityIndicator,
 } from 'react-native';
+import Constants from 'expo-constants';
+import { MaterialIcons } from '@expo/vector-icons';
 import { useRouter } from 'expo-router';
 
 export default function Order() {
+  // Order Status type definition
   type OrderStatus = 'Delivered' | 'Processing' | 'Canceled';
+  
+  // State for selected tab
   const [selectedTab, setSelectedTab] = useState<OrderStatus>('Delivered');
-
   const tabs: OrderStatus[] = ['Delivered', 'Processing', 'Canceled'];
   const router = useRouter();
 
-  const ordersData = {
-    Delivered: [
-      { id: '1', orderNo: 'Order No238562312', orderDate: '21/03/2024', quantity: '03', totalAmount: '$150', status: 'Delivered' },
-      { id: '2', orderNo: 'Order No2385623232', orderDate: '11/03/2022', quantity: '03', totalAmount: '$150', status: 'Delivered' },
-      { id: '3', orderNo: 'Order No2322312', orderDate: '02/13/2021', quantity: '03', totalAmount: '$150', status: 'Delivered' },
-    ],
-    Processing: [
-      { id: '4', orderNo: 'Order No238562312', orderDate: '21/03/2024', quantity: '03', totalAmount: '$150', status: 'Processing' },
-      { id: '5', orderNo: 'Order No2385623232', orderDate: '11/03/2022', quantity: '03', totalAmount: '$150', status: 'Processing' },
-    ],
-    Canceled: [
-      { id: '6', orderNo: 'Order No2322312', orderDate: '02/13/2021', quantity: '03', totalAmount: '$150', status: 'Canceled' },
-    ],
-  };
+  // Replace dummy order data with state that will hold fetched orders grouped by status.
+  const [ordersData, setOrdersData] = useState({
+    Delivered: [] as any[],
+    Processing: [] as any[],
+    Canceled: [] as any[],
+  });
+  const [loading, setLoading] = useState(true);
 
+  // Review Modal related states
+  const [reviewModalVisible, setReviewModalVisible] = useState(false);
+  const [rating, setRating] = useState(0);
+  const [reviewText, setReviewText] = useState('');
+  const [selectedReviewOrder, setSelectedReviewOrder] = useState<any>(null);
+
+  // Use Expo Constants to build API base URL.
+  const API_BASE_URL =
+    Constants.expoConfig?.hostUri?.split(":").shift() ?? "localhost";
+
+  // Fetch orders when the component mounts.
+  useEffect(() => {
+    async function fetchOrders() {
+      try {
+        // For example, fetching orders for the current logged-in user
+        const response = await fetch(`http://${API_BASE_URL}:8000/api/store/orders/my/`);
+        if (!response.ok) {
+          throw new Error('Network response was not ok');
+        }
+        const data = await response.json();
+        
+        // Group orders by status
+        const groupedOrders = {
+          Delivered: [] as any[],
+          Processing: [] as any[],
+          Canceled: [] as any[],
+        };
+        data.forEach((order: any) => {
+          if (order.status === 'Delivered') {
+            groupedOrders.Delivered.push(order);
+          } else if (order.status === 'Processing') {
+            groupedOrders.Processing.push(order);
+          } else if (order.status === 'Canceled') {
+            groupedOrders.Canceled.push(order);
+          }
+        });
+        
+        setOrdersData(groupedOrders);
+      } catch (error) {
+        console.error('Error fetching orders:', error);
+        Alert.alert('Error', 'Failed to fetch orders.');
+      } finally {
+        setLoading(false);
+      }
+    }
+    fetchOrders();
+  }, [API_BASE_URL]);
+
+  // Calculate tab widths and indicator position.
   const tabContainerWidth = Dimensions.get('window').width;
   const tabWidth = tabContainerWidth / tabs.length;
   const indicatorLeft = tabs.indexOf(selectedTab) * tabWidth + (tabWidth * 0.33);
+
+  // ----------------- Review Functionality ----------------- //
+  const handleReview = (order: any) => {
+    setSelectedReviewOrder(order);
+    setRating(0);
+    setReviewText('');
+    setReviewModalVisible(true);
+  };
+
+  const renderStarRating = () => {
+    const stars = [];
+    for (let i = 1; i <= 5; i++) {
+      stars.push(
+        <TouchableOpacity key={i} onPress={() => setRating(i)}>
+          <MaterialIcons
+            name={i <= rating ? 'star' : 'star-border'}
+            size={32}
+            color="gold"
+          />
+        </TouchableOpacity>
+      );
+    }
+    return <View style={styles.starContainer}>{stars}</View>;
+  };
+
+  const submitReview = async () => {
+    if (!selectedReviewOrder) return;
+    if (rating < 1 || rating > 5) {
+      Alert.alert("Please select a rating between 1 and 5.");
+      return;
+    }
+
+    const reviewData = {
+      product: selectedReviewOrder.id,
+      rating,
+      comment: reviewText,
+    };
+
+    console.log("Submitting review payload:", reviewData);
+
+    try {
+      const response = await fetch(`http://${API_BASE_URL}:8000/api/store/reviews/`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(reviewData),
+      });
+
+      if (!response.ok) {
+        const errorData = await response.json();
+        console.error("Failed to submit review", response.statusText, errorData);
+        Alert.alert("Error", "Failed to submit review. Please try again later.");
+        return;
+      }
+
+      console.log("Review submitted successfully.");
+      Alert.alert("Success", "Your review has been submitted.");
+      setReviewModalVisible(false);
+      router.push("/orders");
+    } catch (error) {
+      console.error("Error submitting review:", error);
+      Alert.alert("Error", "An error occurred while submitting your review.");
+    }
+  };
+  // ----------------- End Review Functionality ----------------- //
 
   return (
     <SafeAreaView style={styles.container}>
       <View style={styles.fixedHeader}>
         <View style={styles.header}>
           <TouchableOpacity testID="back-button" onPress={() => router.back()}>
-                      <ImageBackground
-                        style={styles.backIcon}
-                        source={require('@/assets/images/back-arrow.png')}
-                        resizeMode="cover"
-                      />
-                    </TouchableOpacity>
+            <ImageBackground
+              style={styles.backIcon}
+              source={require('@/assets/images/back-arrow.png')}
+              resizeMode="cover"
+            />
+          </TouchableOpacity>
           <Text style={styles.headerTitle} numberOfLines={1}>
             My order
           </Text>
@@ -61,7 +177,9 @@ export default function Order() {
         <View style={styles.orderTabs}>
           {tabs.map(tab => (
             <TouchableOpacity key={tab} onPress={() => setSelectedTab(tab)}>
-              <Text style={selectedTab === tab ? styles.tabActive : styles.tabInactive} numberOfLines={1}>
+              <Text
+                style={selectedTab === tab ? styles.tabActive : styles.tabInactive}
+                numberOfLines={1}>
                 {tab}
               </Text>
             </TouchableOpacity>
@@ -70,58 +188,104 @@ export default function Order() {
         <View style={[styles.tabIndicator, { marginLeft: indicatorLeft }]} />
       </View>
 
-      <ScrollView style={styles.scrollContainer} contentInsetAdjustmentBehavior="automatic">
-        <View style={styles.contentContainer}>
-          {ordersData[selectedTab].map(order => (
-            <View key={order.id} style={styles.orderCard}>
-              <View style={styles.orderCardHeader}>
-                <Text style={styles.orderNo} numberOfLines={1}>
-                  {order.orderNo}
-                </Text>
-                <Text style={styles.orderDate} numberOfLines={1}>
-                  {order.orderDate}
-                </Text>
-              </View>
-              <View style={styles.divider} />
-              <View style={styles.orderCardDetail}>
-                <Text style={styles.totalAmount}>
-                  <Text style={styles.detailLabel}>Quantity:</Text>
-                  <Text style={styles.detailValue}> {order.quantity}</Text>
-                </Text>
-                <Text style={styles.orderTotal}>
-                  <Text style={styles.detailLabel}>Total Amount: </Text>
-                  <Text style={styles.totalAmount}>{order.totalAmount}</Text>
-                </Text>
-              </View>
-              <View style={styles.orderCardFooter}>
-                <View style={styles.detailButton}>
-                  <Text style={styles.detailButtonText} numberOfLines={1}>
-                    Detail
-                  </Text>
+      {loading ? (
+        <ActivityIndicator size="large" color="#232323" style={{ marginTop: 20 }} />
+      ) : (
+        <ScrollView style={styles.scrollContainer} contentInsetAdjustmentBehavior="automatic">
+          <View style={styles.contentContainer}>
+            {ordersData[selectedTab].length > 0 ? (
+              ordersData[selectedTab].map(order => (
+                <View key={order.id} style={styles.orderCard}>
+                  <View style={styles.orderCardHeader}>
+                    <Text style={styles.orderNo} numberOfLines={1}>
+                      {order.orderNo}
+                    </Text>
+                    <Text style={styles.orderDate} numberOfLines={1}>
+                      {order.orderDate}
+                    </Text>
+                  </View>
+                  <View style={styles.divider} />
+                  <View style={styles.orderCardDetail}>
+                    <Text style={styles.totalAmount}>
+                      <Text style={styles.detailLabel}>Quantity:</Text>
+                      <Text style={styles.detailValue}> {order.quantity}</Text>
+                    </Text>
+                    <Text style={styles.orderTotal}>
+                      <Text style={styles.detailLabel}>Total Amount: </Text>
+                      <Text style={styles.totalAmount}>{order.totalAmount}</Text>
+                    </Text>
+                  </View>
+                  <View style={styles.orderCardFooter}>
+                    <View style={styles.buttonGroup}>
+                      <TouchableOpacity style={styles.detailButton}>
+                        <Text style={styles.detailButtonText} numberOfLines={1}>
+                          Detail
+                        </Text>
+                      </TouchableOpacity>
+                      <TouchableOpacity style={styles.reviewButton} onPress={() => handleReview(order)}>
+                        <Text style={styles.detailButtonText} numberOfLines={1}>
+                          Review
+                        </Text>
+                      </TouchableOpacity>
+                    </View>
+                    <Text
+                      style={[
+                        styles.orderStatus,
+                        {
+                          color:
+                            order.status === 'Delivered'
+                              ? '#27ae60'
+                              : order.status === 'Processing'
+                              ? '#F79E1B'
+                              : order.status === 'Canceled'
+                              ? '#EB001B'
+                              : '#27ae60',
+                        },
+                      ]}
+                      numberOfLines={1}>
+                      {order.status}
+                    </Text>
+                  </View>
                 </View>
-                <Text
-                  style={[
-                    styles.orderStatus,
-                    {
-                      color:
-                        order.status === 'Delivered'
-                          ? '#27ae60'
-                          : order.status === 'Processing'
-                          ? '#F79E1B'
-                          : order.status === 'Canceled'
-                          ? '#EB001B'
-                          : '#27ae60',
-                    },
-                  ]}
-                  numberOfLines={1}
-                >
-                  {order.status}
-                </Text>
-              </View>
+              ))
+            ) : (
+              <Text style={{ textAlign: 'center', marginTop: 20 }}>
+                No orders found in this category.
+              </Text>
+            )}
+          </View>
+        </ScrollView>
+      )}
+
+      {/* Review Modal */}
+      <Modal
+        visible={reviewModalVisible}
+        animationType="slide"
+        transparent={true}
+        onRequestClose={() => setReviewModalVisible(false)}>
+        <View style={styles.modalContainer}>
+          <View style={styles.modalContent}>
+            <Text style={styles.modalHeader}>Review Product</Text>
+            {renderStarRating()}
+            <TextInput
+              style={styles.textInput}
+              placeholder="Write your review (max 255 characters)"
+              maxLength={255}
+              multiline
+              value={reviewText}
+              onChangeText={setReviewText}
+            />
+            <View style={styles.modalButtonContainer}>
+              <TouchableOpacity style={styles.submitButton} onPress={submitReview}>
+                <Text style={styles.buttonText}>Submit Review</Text>
+              </TouchableOpacity>
+              <TouchableOpacity style={styles.cancelButton} onPress={() => setReviewModalVisible(false)}>
+                <Text style={styles.buttonText}>Cancel</Text>
+              </TouchableOpacity>
             </View>
-          ))}
+          </View>
         </View>
-      </ScrollView>
+      </Modal>
     </SafeAreaView>
   );
 }
@@ -291,12 +455,26 @@ const styles = StyleSheet.create({
     paddingHorizontal: 21,
     height: 36,
   },
+  buttonGroup: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
   detailButton: {
     width: 100,
     height: 36,
     backgroundColor: '#232323',
     borderRadius: 4,
     justifyContent: 'center',
+    alignItems: 'center',
+  },
+  reviewButton: {
+    width: 100,
+    height: 36,
+    backgroundColor: '#232323',
+    borderRadius: 4,
+    justifyContent: 'center',
+    alignItems: 'center',
+    marginLeft: 8,
   },
   detailButtonText: {
     fontFamily: 'Nunito Sans',
@@ -311,7 +489,65 @@ const styles = StyleSheet.create({
     fontSize: 16,
     fontWeight: '600',
     lineHeight: 21.824,
-    color: '#27ae60',
     textAlign: 'right',
+  },
+  // ----------------- Review Modal Styles ----------------- //
+  modalContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0,0,0,0.5)',
+  },
+  modalContent: {
+    width: '90%',
+    backgroundColor: '#fff',
+    borderRadius: 10,
+    padding: 20,
+    alignItems: 'center',
+  },
+  modalHeader: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginBottom: 15,
+  },
+  starContainer: {
+    flexDirection: 'row',
+    marginBottom: 15,
+  },
+  textInput: {
+    width: '100%',
+    height: 100,
+    borderColor: '#ccc',
+    borderWidth: 1,
+    borderRadius: 5,
+    padding: 10,
+    marginBottom: 15,
+    textAlignVertical: 'top',
+  },
+  modalButtonContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    width: '100%',
+  },
+  submitButton: {
+    backgroundColor: 'green',
+    paddingVertical: 10,
+    paddingHorizontal: 15,
+    borderRadius: 5,
+    flex: 1,
+    marginRight: 5,
+  },
+  cancelButton: {
+    backgroundColor: 'gray',
+    paddingVertical: 10,
+    paddingHorizontal: 15,
+    borderRadius: 5,
+    flex: 1,
+    marginLeft: 5,
+  },
+  buttonText: {
+    color: '#fff',
+    fontWeight: 'bold',
+    textAlign: 'center',
   },
 });
