+++ mobile/package.json
@@ -60,8 +60,7 @@
     "react-native-toast-message": "^2.2.1",
     "react-native-vector-icons": "^10.2.0",
     "react-native-web": "~0.19.13",
-    "react-native-webview": "13.12.5",
-    "react-native-maps": "1.18.0"
+    "react-native-webview": "13.12.5"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
@@ -76,6 +75,7 @@
     "@types/react-native-vector-icons": "^6.4.18",
     "@types/react-navigation": "^3.4.0",
     "@types/react-test-renderer": "^18.3.0",
+    "babel-plugin-module-resolver": "^5.0.2",
     "expo-module-scripts": "^4.0.4",
     "jest": "^29.7.0",
     "jest-expo": "~52.0.3",
